edit

+++ printf/ft_print_int.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/13 22:01:35 by hkubo             #+#    #+#             */
-/*   Updated: 2021/03/14 11:22:31 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/03/14 11:22:38 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int print_int(va_list *ap, a_list flag_list)
     len = 0;
     num = va_arg(*ap, int);
     str_num = ft_itoa(num);
-    if (flag_list.field > ft_strlen(str_num))
+    if (flag_list.field > (int)ft_strlen(str_num))
         printf("ok");
     printf("num: %d\n", num);
     printf("flag: %d\n", flag_list.precision);
